% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_meta_data.R
\name{extract_meta_data}
\alias{extract_meta_data}
\title{Extract meta data}
\usage{
extract_meta_data(html_doc)
}
\arguments{
\item{html_doc}{Charcter string including the html document.}
}
\value{
List with three sublists for native, open graph and twitter.
}
\description{
\code{extract_meta_data} extracts native, open graph and twitter meta data
from html documents. The meta data include url, title, description and image.
The html document is parsed within the function
}
