% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tags_id.R, R/get_tag_id.R
\name{find_tags_id}
\alias{find_tags_id}
\title{Find Tags ID Given A Tag Sets ID}
\usage{
find_tags_id(tag_name, tag_sets_id, last_tags_id = 0, rows = 20)

find_tags_id(tag_name, tag_sets_id, last_tags_id = 0, rows = 20)
}
\arguments{
\item{tag_name}{character, the name of the tag for which the id should be found.}

\item{tag_sets_id}{character, the tag sets id of the tag sets to search in.}

\item{last_tags_id}{integer, Return tags with a tags_id is greater than this value (default = 0)}

\item{rows}{integer, number of tag sets to return. Cannot be larger than 100 (default = 20)}

\item{tag_name}{character, the name of the tag for which the id should be found.}

\item{tag_sets_id}{character, the tag sets id of the tag sets to search in.}

\item{last_tags_id}{integer, Return tags with a tags_id is greater than this value (default = 0)}

\item{rows}{integer, number of tag sets to return. Cannot be larger than 100 (default = 20)}
}
\value{
...

...
}
\description{
The function finds the tags_id given the name of a tag and a tag_sets_id.

The function finds the tags_id given the name of a tag and a tag_sets_id.
}
\details{
See the example "Find the tags_id for 'odd' given the tag_sets_id" in the official API documentation. 
This is an R implementation of essentially the same function.

See the example "Find the tags_id for 'odd' given the tag_sets_id" in the official API documentation. 
This is an R implementation of essentially the same function.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# 
\dontrun{tagsids <- find_tags_id(tag_name = "newspapers", tag_sets_id = 5, rows = 2)}
\dontrun{tagsids}
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# 
\dontrun{tagsids <- find_tags_id(tag_name = "newspapers", tag_sets_id = 5, rows = 2)}
\dontrun{tagsids}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md\cr

https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>

Ulrich Matter <umatter-at-protonmail.com>
}
