% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags_id.R
\name{get_tags_id}
\alias{get_tags_id}
\title{Get Tag IDs}
\usage{
get_tags_id(tag_name, exact=FALSE, source="local")
}
\arguments{
\item{tag_name}{character, the name of the tag for which the id should be found.}

\item{source}{character, either "local" or "online". if "online" the tags list is downloaded from mediacloud when this function is called.
if "local" a local copy of the list (might not be up to date) shipped with mediacloudr is used. (default = "local")}

\item{fuzzy}{logical, if TRUE, only exact matches are returned. If FALSE tab_name is considered a regular expression. (default = FALSE)}
}
\value{
a data frame containing the found tag names and tag ids
}
\description{
The function returns specific tag ids and tag names from the nyt-theme tag list provided by MediaCloud
}
\details{
...
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# 
\dontrun{govtag_id <- get_tags_id(tag_name = "united states politics")}
\dontrun{govtag_id}
}
\references{
https://mediacloud.org/support/theme-list
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
