% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_themes_list.R
\name{get_themes_list}
\alias{get_themes_list}
\title{Return List of Themes}
\usage{
get_themes_list()
}
\value{
a data.table object
}
\description{
Fetch theme list from https://mediacloud.org/s/nyt-theme-tags.csv as data table.
}
\details{
...
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# https://api.mediacloud.org/api/v2/media/list?last_media_id=1&rows=2

\dontrun{get_themes_list()}
}
\references{
https://mediacloud.org/support/theme-list
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
