% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_storiescount_mediaids.R
\name{grab_storiescount_mediaids}
\alias{grab_storiescount_mediaids}
\title{Return All Counts of stories Associated with a Media ID}
\usage{
grab_storiescount_mediaids(media_id, fq = NULL, split = NULL, split_period = "day")
}
\arguments{
\item{media_id}{character vector, the media ids of the outlets for which all counts shall be fetched.}

\item{fq}{("filter query") parameter which the API method directly passes to Solr (default = NULL)}

\item{split}{if set to 1 or true, split the counts into date ranges (default = NULL)}

\item{split_period}{character, indicating for which date periods counts should be returned: day, week, month, year (default = "day")}

\item{max_calls}{integer, number of total calls to the API. If NULL, no restrictions are applied, all counts will be collected. (default=NULL)}
}
\value{
...
}
\description{
Fetch data on all story counts corresponding to a given media id via the api/v2/counts_public/count method of the MediaCloud API.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# fetch count data on stories about US politics in the us mainstream media for the time frame 2014-01-01T00:00:00Z TO 2014-03-01T00:00:00Z
\dontrun{storiescount <- grab_storiescount_mediatag(tags_id_media="8875027", fq="tags_id_stories:9360846 AND publish_day:[2014-01-01T00:00:00Z TO 2014-03-01T00:00:00Z]")}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2counts_publiccount\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
