% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_health_list.R
\name{media_health_list}
\alias{media_health_list}
\title{Return Media Sources Health}
\usage{
media_health_list(media_id)
}
\arguments{
\item{media_id}{integer, media id of sources to fetch data on}
}
\value{
A list of two data frames, one with the general health variables, and one on the coverage gaps
}
\description{
Fetch data from the api/v2/mediahealth/list method of the MediaCloud API.
}
\details{
The Media Health API call provides information about the health of a media source, 
meaning to what degree we are capturing all of the stories published by that media source. 
Media Cloud collects its data via automatically detected RSS feeds on the open web. 
This means first that the system generally has data for a given media source from the time we 
first enter that source into our database. Second, Media Cloud data for a given media source is 
only as good as the set of feeds we have for that source. Our feed scraper is not perfect and so 
sometimes misses feeds it should be collecting. Third, feeds change over time. We periodically 
rescrape every media source for new feeds, but this takes time and is not perfect.
The only way we have of judging the health is judging the relative number of stories over time. 
This media call provides a set of metrics that compare the current number of stories being 
collected by the media source with the number of stories collected over the past 90 days, 
and also compares coverage over time with the expected volume. More details are in the field 
descriptions in the reference link to the github documentation of the API.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}

\dontrun{media_health_list(media_id =1)}
}
\references{
https://github.com/mediacloud/backend/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2mediahealthlist
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
