\name{mediacloudr-package}
\alias{mediacloudr-package}
\alias{mediacloudr}
\docType{package}
\title{
\packageTitle{mediacloudr}
}
\description{
\packageDescription{mediacloudr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mediacloudr}
\packageIndices{mediacloudr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mediacloudr}

Maintainer: \packageMaintainer{mediacloudr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
