% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data_html.R
\docType{data}
\name{meta_data_html}
\alias{meta_data_html}
\title{HTML document to test \code{extract_meta_data}}
\format{An object of class \code{character} of length 1.}
\usage{
meta_data_html
}
\description{
A HTML document with basic meta tags for open-graph, twitter and native
meta data.
}
\keyword{datasets}
