% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories_public_count.R
\name{stories_public_count}
\alias{stories_public_count}
\title{Stories Public Count}
\usage{
stories_public_count(q, fq = NULL, split = NULL, split_period = "day")
}
\arguments{
\item{q}{("query") parameter which the API method directly passes to Solr}

\item{fq}{("filter query") parameter which the API method directly passes to Solr (default = NULL)}

\item{split}{if set to 1 or true, split the counts into date ranges (default = NULL)}

\item{split_period}{character, indicating for which date periods counts should be returned: day, week, month, year (default = "day")}
}
\value{
...
}
\description{
Fetch data from the api/v2/stories_public/count method of the MediaCloud API.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# fetch count of stories containing the word 'obama' in the New York Times 
# (https://api.mediacloud.org/api/v2/stories_public/count?q=obama&fq=media_id:1)
\dontrun{obama <- stories_public_count(q="obama", fq="media_id=1")}
\dontrun{obama}
# fetch count of stories containing 'africa' in the New York Times for each week from 2014-01-01 to 2014-03-01:
\dontrun{africa <- stories_public_count(q="africa AND media_id:1", fq="publish_day:[2014-01-01T00:00:00Z TO 2014-03-01T00:00:00Z]", split=1, split_period = "week")}
\dontrun{africa}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2stories_publiccount\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
