% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories_public_list.R
\name{stories_public_list}
\alias{stories_public_list}
\title{Return Multiple Processed Stories}
\usage{
stories_public_list(last_processed_stories_id = 0, rows = 20, feeds_id = NULL, q=NULL, fq = NULL, sort="processed_stories_id", wc=0, show_feeds=0)
}
\arguments{
\item{last_processed_stories_id}{integer, Return stories in which the processed_stories_id is greater than this value. (default = 0)}

\item{rows}{integer, number of tag sets to return. Cannot be larger than 100 (default:20)}

\item{feeds_id}{character, Return only stories that match the given feeds_id, sorted my descending publish date (default = NULL)}

\item{q}{("query") parameter which the API method directly passes to Solr (default = NULL)}

\item{fq}{("filter query") parameter which the API method directly passes to Solr (default = NULL)}

\item{sort}{character, Returned results sort order. Supported values: processed_stories_id, random (default = processed_stories_id)}

\item{wc}{integer, if set to 1, include a 'word_count' field with each story that includes a count of the most common words in the story (default = 0)}

\item{show_feeds}{integer, if set to 1, include a 'feeds' field with a list of the feeds associated with this story (default=0)}
}
\value{
...
}
\description{
Fetch data from the api/v2/stories_public/list method of the MediaCloud API.
}
\details{
The last_processed_stories_id parameter can be used to page through these results. The API will return stories with aprocessed_stories_id greater than this value.
To get a continuous stream of stories as they are processed by Media Cloud, the user must make a series of calls to api/v2/stories_public/list in which last_processed_stories_id for
each call is set to the processed_stories_id of the last story in the previous call to the API. A single call can only return up to 10,000 results, but you can get the full list of results by paging through the full list using last_processed_stories_id. 
Note: stories_id and processed_stories_id are separate values. The order in which stories are processed is different than the stories_id order. The processing pipeline involves downloading, extracting, and vectoring stories. Requesting by the processed_stories_id field guarantees that the user will receive every story (matching the query criteria if present) in the order it is processed by the system.
The q and fq parameters specify queries to be sent to a Solr server that indexes all Media Cloud stories. The Solr server provides full text search indexing of each sentence collected by Media Cloud. All content is stored as individual sentences. The api/v2/stories_public/list call searches for sentences matching the q and / or fq parameters if specified and the stories that include at least one sentence returned by the specified query.
The q and fq parameters are passed directly through to Solr.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# Return a stream of all stories from The New York Times mentioning 'obama' greater than the given last_processed_stories_id.
# (https://api.mediacloud.org/api/v2/stories_public/list?last_processed_stories_id=2523432&q=text:obama+AND+media_id:1)
\dontrun{obama_tags <- stories_public_list(last_processed_stories_id ="2523432", q="obama AND media_id:1")}
\dontrun{obama_tags}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2stories_publiclist\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
