% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories_public_tagcount.R
\name{stories_public_tagcount}
\alias{stories_public_tagcount}
\title{Stories Public Tag Count}
\usage{
stories_public_tagcount(q, fq = NULL, limit=1000, tag_sets_id=NULL)
}
\arguments{
\item{q}{("query") parameter which the API method directly passes to Solr}

\item{fq}{("filter query") parameter which the API method directly passes to Solr (default = NULL)}

\item{limit}{numeric, number of tags to fetch from Solr (default = 1000)}

\item{tag_sets_id}{character, return only tags belonging to this tag set (default = NULL)}
}
\value{
...
}
\description{
Fetch data from the api/v2/stories_public/tab_count method of the MediaCloud API.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# fetch Count pf tags in stories containing the word 'obama' in The New York Times. 
# (https://api.mediacloud.org/api/v2/stories_public/tag_count?q=obama&fq=media_id:1&limit=3)
\dontrun{obama_tags <- stories_public_tagcount(q="obama", fq="media_id=1", limit=3)}
\dontrun{obama_tags}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2stories_publiccount\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
