% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_sets_list.R
\name{tag_sets_list}
\alias{tag_sets_list}
\title{Stories Public Tag Count}
\usage{
tag_sets_list(last_tag_sets_id = 0, rows = 20)
}
\arguments{
\item{last_tag_sets_id}{integer, return tag sets with a tag_sets_id greater than this value (default = 0)}

\item{rows}{integer, number of tag sets to return. Cannot be larger than 100 (default:20)}
}
\value{
...
}
\description{
Fetch data from the api/v2/stories_public/tab_count method of the MediaCloud API.
}
\examples{
# First, make sure your personal Mediacloud API key is saved as character string in the mc.key variable:
\dontrun{mc.key <- "yourkey"}
# https://api.mediacloud.org/api/v2/tag_sets/list
\dontrun{tagslist <- tag_sets_list()}
\dontrun{tagslist}
}
\references{
https://github.com/berkmancenter/mediacloud/blob/master/doc/api_2_0_spec/api_2_0_spec.md#apiv2stories_publiccount\cr
}
\author{
Ulrich Matter <umatter-at-protonmail.com>
}
